♻️ Post seperation
- Post 컴포넌트 분리
