+++ packages/client/src/entities/posts/index.ts
@@ -1,0 +1,1 @@
+export { default as Posts } from './ui/Posts';
