+++ packages/client/src/shared/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useFetch';
+export * from './useForwardRef';
